--- src/tailwind/v3/index.ts
@@ -1,1 +1,0 @@
-export { getClassOrder } from "./class-order.js";
